Authentication, file upload, post image display
Implemented user Authentication, File upload and viewing of post images

+++ app/Http/Controllers/PostsController.php
@@ -9,6 +9,16 @@ use App\Post;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -44,13 +54,36 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        // Handle File Upload
+        if($request->hasFile('cover_image')){
+            // Retrieve file name with extention
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+
+            // Retrieve file name
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME); 
+            
+            // Retrieve file extension 
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            
+            // File to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            
+            // Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        }else{
+            $fileNameToStore =  'noimage.jpg';
+        }
+
         // Create Post 
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body'); 
-        $post->user_id = auth()->user()->id;   
+        $post->user_id = auth()->user()->id; 
+        $post->cover_image = $fileNameToStore;  
         $post->save();
         
         // Redirect post with success messaging
@@ -78,6 +111,12 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        // Check for correct user
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts ')->with('error', 'Unauthorized Page');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -94,10 +133,32 @@ class PostsController extends Controller
             'title' => 'required',
             'body' => 'required'
         ]);
+
+        // Handle File Upload
+        if($request->hasFile('cover_image')){
+            // Retrieve file name with extention
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+
+            // Retrieve file name
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME); 
+            
+            // Retrieve file extension 
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            
+            // File to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            
+            // Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        }
+
         // Create Post 
         $post = Post::find($id);
         $post->title = $request->input('title');
-        $post->body = $request->input('body');    
+        $post->body = $request->input('body'); 
+        if($request->hasFile('cover_image')){
+            $post->cover_image = $fileNameToStore; 
+        }
         $post->save();
         
         // Redirect post with success messaging
@@ -113,6 +174,15 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        // Check for correct user
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts ')->with('error', 'Unauthorized Page');
+        }
+        if($post->cover_image != 'noimage/jpg'){
+            //Delete Image
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
         $post->delete();
         return redirect('/posts')->with('success', 'Post Removed');
     }
